Break apart bullet logic and satisfy more typescript types.
